% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bearer_header.R
\name{bearer_header}
\alias{bearer_header}
\title{Construct bearer header}
\usage{
bearer_header(access_token = jfrog_access_token())
}
\arguments{
\item{access_token}{A valid token.}
}
\value{
A vector that can be used in a header.
}
\description{
Construct bearer header for authorization with with an access token.
}
\examples{
\dontrun{
install.packages("<package>", headers = bearer_header())
}

}
