% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_token.R
\name{create_token}
\alias{create_token}
\title{Create new access token}
\usage{
create_token(
  jfrog_url,
  expires_in = 2 * 60,
  access_token = jfrog_access_token()
)
}
\arguments{
\item{jfrog_url}{Base URL of the JFrog domain.}

\item{expires_in}{The life span of the token in seconds. Defaults to 2 minutes.}

\item{access_token}{A valid token.}
}
\value{
A parsed response from the token endpoint.
}
\description{
Get a token from the endpoint "access/api/v1/tokens" per \href{https://www.jfrog.com/confluence/display/JFROG/Artifactory+REST+API#ArtifactoryRESTAPI-AccessTokens}{JFrog's documentation}.
Note that the documentation for "expiry" parameter is \emph{wrong} at the time of writing.
}
\details{
A token can only be created using already existing tokens.
A usecase is to have a long-lived token to generate more transient token used for authentication as part of e.g. test pipelines.
}
