% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jfrog-package.R
\docType{package}
\name{jfrog-package}
\alias{jfrog}
\alias{jfrog-package}
\title{jfrog package}
\description{
The {jfrog} package provides an interface to JFrog's CRAN and artifactory.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/robertdj/jfrog}
}

}
\author{
\strong{Maintainer}: Robert Dahl Jacobsen \email{git@dahl-jacobsen.dk}

}
