% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jfrog_access_token.R
\name{jfrog_access_token}
\alias{jfrog_access_token}
\title{Retrieve JFrog access token}
\usage{
jfrog_access_token(quiet = TRUE)
}
\arguments{
\item{quiet}{Whether or not to display a message if the access token is \emph{not} available.}
}
\value{
The access token if available and \code{NA} otherwise.
}
\description{
Retrieve a JFrog access token.
Looks in the environment variable \code{JFROG_ACCESS_TOKEN}.
}
