% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jfrog_api_key.R
\name{jfrog_api}
\alias{jfrog_api}
\title{Retrieve JFrog API Key}
\usage{
jfrog_api(quiet = TRUE)
}
\arguments{
\item{quiet}{Whether or not to display a message if the API Key is \emph{not} available.}
}
\value{
The API Key if available and \code{NA} otherwise.
}
\description{
Retrieve a JFrog API Key.
Looks in the environment variable \code{JFROG_API_KEY}.
}
