% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upload_package.R
\name{upload_package}
\alias{upload_package}
\title{Upload package archive to JFrog}
\usage{
upload_package(package_archive, jfrog_url, api_key = jfrog_pat(), access_token)
}
\arguments{
\item{package_archive}{Path to \code{tar.gz}, \code{zip} or \code{tgz} archive.}

\item{jfrog_url}{URL}

\item{api_key}{JFrog API Key.}

\item{access_token}{JFrog access token.}
}
\value{
Response from \code{\link[httr:POST]{httr::POST()}}.
}
\description{
JFrog allows different ways of authentication as noted in \href{https://www.jfrog.com/confluence/display/JFROG/Artifactory+REST+API}{JFrog's documentation}.
Here we rely on an \href{https://www.jfrog.com/confluence/display/JFROG/User+Profile#UserProfile-APIKey}{API Key} for a dedicated header or an \href{https://www.jfrog.com/confluence/display/JFROG/Access+Tokens}{access token} in an authorization header.
}
\details{
If both \code{api_key} and \code{access_token} are supplied then \code{access_token} is used.
The rationale is that this is the desired behavior in CI pipelines.
}
